pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-table-engine"
version = "0.1.0"
description = "Production-ready LLM-assisted table ingestion engine"
authors = [{name = "Facturio"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.29.0",
    "pydantic==2.6.4",
    "pydantic-settings==2.2.1",
    "structlog==24.1.0",
    "pandas==2.2.1",
    "python-magic==0.4.27",
    "chardet==5.2.0",
    "openpyxl==3.1.2",
    "xlrd==2.0.1",
    "gspread==6.0.2",
    "google-auth==2.27.0",
    "google-cloud-bigquery==3.17.1",
    "pandas-gbq==0.21.0",
    "python-dateutil==2.9.0",
    "unidecode==1.3.8",
    "httpx==0.27.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.3.5",
    "black==24.2.0",
    "pytest==8.1.1",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
